requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
flask-restful
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
alembic
pytest
python-dotenv
googlemaps

